versio with payment process 40% complete
