+++ Homiegram_front_end/lib/views/Shared/house_details_screen.dart
@@ -4,7 +4,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_rating_bar/flutter_rating_bar.dart';
 import 'package:homi_2/components/blured_image.dart';
 import 'package:homi_2/components/my_snackbar.dart';
-import 'package:homi_2/components/not_found.dart';
 import 'package:homi_2/models/amenities.dart';
 import 'package:homi_2/models/bookmark.dart';
 import 'package:homi_2/models/comments.dart';
@@ -16,7 +15,6 @@ import 'package:homi_2/services/fetch_bookmarks.dart';
 import 'package:homi_2/services/get_amenities.dart';
 import 'package:homi_2/services/get_house_service.dart';
 import 'package:homi_2/services/get_locations.dart';
-import 'package:homi_2/services/rent_room_service.dart';
 import 'package:homi_2/services/user_data.dart';
 import 'package:homi_2/services/user_sigin_service.dart';
 import 'package:homi_2/views/Shared/comments_screen.dart';
