+++ Homiegram_front_end/lib/views/Tenants/renting_page.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:homi_2/components/action_button.dart';
-import 'package:homi_2/models/room.dart';
 import 'package:homi_2/models/room_with_agrrement_model.dart';
 import 'package:homi_2/services/get_rooms_service.dart';
 import 'package:homi_2/services/user_data.dart';
@@ -221,8 +220,10 @@ class _RentingPageState extends State<RentingPage> {
                                       ),
                                     ),
                                   ),
-                                  const SizedBox(height: 12),
+                                  const SizedBox(height: 6),
                                   Row(
+                                    mainAxisAlignment:
+                                        MainAxisAlignment.spaceEvenly,
                                     children: [
                                       ActionButton(
                                         label: "Raise Complaint",
@@ -231,15 +232,13 @@ class _RentingPageState extends State<RentingPage> {
                                             const Color(0xFFF0B803),
                                         onPressed: () {},
                                       ),
-                                      const SizedBox(width: 12),
+                                      const SizedBox(width: 20),
                                       ActionButton(
-                                        label: "Terminate",
-                                        icon: Icons.cancel,
+                                        label: "Pay Rent",
+                                        icon: Icons.money,
                                         backgroundColor:
-                                            const Color(0xFF940B01),
-                                        onPressed: () {
-                                          _confirmTermination(context);
-                                        },
+                                            const Color(0xFF126E06),
+                                        onPressed: () {},
                                       ),
                                     ],
                                   ),
@@ -283,41 +282,6 @@ class _RentingPageState extends State<RentingPage> {
             ),
           ],
         ),
-      ),
-    );
-  }
-
-  void _confirmTermination(BuildContext context) {
-    showDialog(
-      context: context,
-      builder: (_) => AlertDialog(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(20)),
-        title: const Text(
-          "Confirm Termination",
-          style: TextStyle(fontWeight: FontWeight.bold),
-        ),
-        content: const Text(
-            "Are you sure you want to terminate your contract? This cannot be undone."),
-        actions: [
-          TextButton(
-            onPressed: () => Navigator.pop(context),
-            child: const Text(
-              "Cancel",
-              style: TextStyle(color: Colors.black54),
-            ),
-          ),
-          ElevatedButton(
-            style: ElevatedButton.styleFrom(
-              backgroundColor: const Color(0xFF940B01),
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(12),
-              ),
-            ),
-            onPressed: () => Navigator.pop(context),
-            child:
-                const Text("Continue", style: TextStyle(color: Colors.white)),
-          ),
-        ],
       ),
     );
   }
